Added a check for today
Encountered an issue where etherscan found a sale on OpenSea, but OpenSea API was malfunctioning and returned a sale that happened days ago. 
Added a check for the event that comes back from OpenSea that makes sure the most recent event happened today. 

Potential errors: if the same NFT was sold multiple times in one day, and if the OpenSea API encounters this same issue, then a sale from earlier on the same day could be tweeted.